google_sheets_mcp/requirements.txt
# MCP SDK
modelcontextprotocol>=0.1.0

# Google API dependencies
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Utility dependencies
python-dotenv>=1.0.0
aiohttp>=3.8.5
pydantic>=2.4.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.1
pylint>=2.17.5
